move code to new additional mainwindow file
